Project divided in modules
