+++ actions.py
@@ -1,31 +1,3 @@
-import csv
-
-
-def get_menu_value():
-    selection = 0
-    print(
-        """
-        Student Control Panel
-
-            1) Add Student
-            2) See all students data
-            3) Top 3 students
-            4) All students average
-            5) Export student data to csv
-            6) Import data from csv
-        """
-    )
-    while True:
-        try:
-            selection = int(input('Select an option from the menu "Press 7 to exit": '))
-            if selection < 1 or selection > 7:
-                raise ValueError()
-            break
-        except ValueError as ex:
-            print ('Invalid selection')
-    return selection
-
-
 def get_student_note(text):
     note = 0
     while True:
@@ -105,7 +77,6 @@ def get_top_student(list, student_number):
                         top_student.append(student_average)
                 if len(top_student) > student_number:
                     top_student.pop()
-
     except IndexError:
         print('No more student in the data base: Index Error')
     except KeyError:
@@ -133,56 +104,3 @@ def show_all_students_average(list):
     except ZeroDivisionError:
         print('No students registered: ZeroDivisionError')
     print(f'All students average: {total_average}')
-
-
-def export_data(list):
-    headers = []
-    try:
-        for key in list[0]:
-            headers.append(key)
-    except IndexError as ex:
-        print('Incorrect list provided')
-    with open('student_list.csv', 'w', encoding='utf-8') as file:
-        writer = csv.DictWriter(file, headers)
-        writer.writeheader()
-        writer.writerows(list)
-
-
-def import_data():
-    student_list = []
-    try:
-        with open('student_list.csv', 'r', encoding='utf-8') as file:
-            reader = csv.DictReader(file)
-            for row in reader:
-                for key in row:
-                    if key[len(key)-4:len(key)] == 'note':
-                        row[key] = int(row[key])
-                student_list.append(row)
-    except FileNotFoundError:
-        print('Document is not available')
-    return student_list
-
-
-def main():
-    student_list = []
-    operator = 0
-    while True:
-        operator = get_menu_value()
-        if operator == 1:
-            student_list.extend(create_student_list())
-        elif operator == 2:
-            show_student_data(student_list)
-        elif operator == 3:
-            show_top_students(student_list)
-        elif operator == 4:
-            show_all_students_average(student_list)
-        elif operator == 5:
-            export_data(student_list)
-        elif operator == 6:
-            student_list = import_data()
-        elif operator == 7:
-            break
-
-
-if __name__=='__main__':
-    main()
